organized code, and updated readme
