+++ README.md
@@ -31,6 +31,7 @@ To solve this question, the following steps are necessary:
 
 ![Solution Schema](graphs\final_graph_schema.jpeg)
 
+
 ## Algorithm Represented by the Graph:
 
 1. Start with the user's question.<br>
@@ -53,6 +54,8 @@ To solve this question, the following steps are necessary:
 8. Verify that the final answer is grounded in the context. If not, generate the final answer again.<br>
 9. Output the final answer to the user.<br>
 
+![Demo](graphs\demo_gif.gif)
+
 ## Heuristics and Techniques Implemented in This Solution:
 
 1. Encoding both book content in chunks and chapter summaries generated by LLM.<br>
@@ -67,6 +70,8 @@ To solve this question, the following steps are necessary:
 ## Prerequisites
 
 - Python 3.8+
+
+## Installation
 
 ## Installation
 
@@ -86,6 +91,12 @@ To solve this question, the following steps are necessary:
     OPENAI_API_KEY=your_openai_api_key
     GROQ_API_KEY=your_groq_api_key
     ```
+
+4. **Step-by-Step Tutorial Notebook:** The step-by-step tutorial notebook is `sophisticated_rag_agent_harry_potter.ipynb`.
+
+5. **Run a Real-Time Visualization of the Agent:**
+    - Open CMD and type: `streamlit run simulate_agent.py`
+
 
 ## Contributing
